refactor: LottoMachineEntities 디렉토리 생성
- LottoMachine에서 쓰이는 Ticket과 WinningCount 클래스를
  LottoMachineEntities에 담음
- 이는 폴더 구조를 통해 Lotto, BonusNumber, LottoMachine이라는
  세개의 주요 객체가 있고, 그 외의 도와주는 객체와 함수들이 있다는 것을
  명확하게 보여주기 위함
